requirements/base.txt
# requirements/base.txt
pandas>=1.3.0
openpyxl>=3.0.0
pyyaml>=5.4.0
click>=8.0.0
networkx>=2.6.0
pydantic>=1.8.0
numpy>=1.20.0